+++ pkg/pkg.ml
@@ -3,9 +3,12 @@
 #require "topkg"
 open Topkg
 
+let cmdliner = Conf.with_pkg "cmdliner"
+
 let () =
   Pkg.describe "otfm" @@ fun c ->
+  let cmdliner = Conf.value c cmdliner in
   Ok [ Pkg.mllib "src/otfm.mllib";
-       Pkg.bin "test/otftrip";
+       Pkg.bin ~cond:cmdliner "test/otftrip";
        Pkg.doc "test/examples.ml";
        Pkg.test "test/examples"; ]